Added writting of a 2 input 1 output gate. Added XOR gate for a single Wire example.

+++ include/Wire.hpp
@@ -8,7 +8,7 @@ namespace gabe {
             class Wire
             {
             public:
-                uint8_t label = 0;
+                uint64_t label = 0;
             };
         }
     }
